requirements.txt
asgiref==3.8.1
Django==5.1.4
django-phonenumber-field==8.0.0
phonenumberslite==8.13.52
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
